printf

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yfrancoi <yfrancoi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/29 17:10:19 by yfrancoi          #+#    #+#             */
-/*   Updated: 2021/11/29 17:17:34 by yfrancoi         ###   ########lyon.fr   */
+/*   Updated: 2021/11/30 17:59:01 by yfrancoi         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include "libft.h"
 # include <stdarg.h>
 
-int	ft_putnbr_base(int nb, char *base);
+int	ft_putnbr_base(long nb, char *base);
 int	ft_printf(const char *format,...);
 
 #endif
